# tp-final/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pulso API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/consumos": {
      "get": {
        "tags": [
          "consumos"
        ],
        "summary": "Index",
        "operationId": "index_api_consumos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsumoListado"
                  },
                  "type": "array",
                  "title": "Response Index Api Consumos Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "consumos"
        ],
        "summary": "Create",
        "operationId": "create_api_consumos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumoIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/consumos/{identifier}": {
      "get": {
        "tags": [
          "consumos"
        ],
        "summary": "Detail",
        "operationId": "detail_api_consumos__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "consumos"
        ],
        "summary": "Update",
        "operationId": "update_api_consumos__identifier__put",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumoIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "consumos"
        ],
        "summary": "Remove",
        "operationId": "remove_api_consumos__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recursos": {
      "get": {
        "tags": [
          "recursos"
        ],
        "summary": "Index",
        "operationId": "index_api_recursos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsuarioOut"
                  },
                  "type": "array",
                  "title": "Response Index Api Recursos Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "recursos"
        ],
        "summary": "Create",
        "operationId": "create_api_recursos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecursoIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recursos/{identifier}": {
      "get": {
        "tags": [
          "recursos"
        ],
        "summary": "Detail",
        "operationId": "detail_api_recursos__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "recursos"
        ],
        "summary": "Update",
        "operationId": "update_api_recursos__identifier__put",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecursoIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "recursos"
        ],
        "summary": "Remove",
        "operationId": "remove_api_recursos__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsumoIn": {
        "properties": {
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "proyecto_id": {
            "type": "integer",
            "title": "Proyecto Id"
          },
          "recurso_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurso Id"
          },
          "horas_consumidas": {
            "type": "number",
            "title": "Horas Consumidas"
          },
          "tarea": {
            "type": "string",
            "title": "Tarea"
          },
          "rol_id": {
            "type": "integer",
            "title": "Rol Id"
          }
        },
        "type": "object",
        "required": [
          "fecha_inicio",
          "fecha_fin",
          "proyecto_id",
          "horas_consumidas",
          "tarea",
          "rol_id"
        ],
        "title": "ConsumoIn"
      },
      "ConsumoListado": {
        "properties": {
          "consumo_id": {
            "type": "integer",
            "title": "Consumo Id"
          },
          "proyecto_id": {
            "type": "integer",
            "title": "Proyecto Id"
          },
          "recurso_id": {
            "type": "integer",
            "title": "Recurso Id"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "horas_consumidas": {
            "type": "number",
            "title": "Horas Consumidas"
          },
          "tarea": {
            "type": "string",
            "title": "Tarea"
          },
          "rol_id": {
            "type": "integer",
            "title": "Rol Id"
          },
          "proyecto_nombre": {
            "type": "string",
            "title": "Proyecto Nombre"
          },
          "recurso_nombre": {
            "type": "string",
            "title": "Recurso Nombre"
          },
          "rol_descripcion": {
            "type": "string",
            "title": "Rol Descripcion"
          }
        },
        "type": "object",
        "required": [
          "consumo_id",
          "proyecto_id",
          "recurso_id",
          "fecha_inicio",
          "fecha_fin",
          "horas_consumidas",
          "tarea",
          "rol_id",
          "proyecto_nombre",
          "recurso_nombre",
          "rol_descripcion"
        ],
        "title": "ConsumoListado"
      },
      "ConsumoOut": {
        "properties": {
          "consumo_id": {
            "type": "integer",
            "title": "Consumo Id"
          },
          "proyecto_id": {
            "type": "integer",
            "title": "Proyecto Id"
          },
          "recurso_id": {
            "type": "integer",
            "title": "Recurso Id"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "horas_consumidas": {
            "type": "number",
            "title": "Horas Consumidas"
          },
          "tarea": {
            "type": "string",
            "title": "Tarea"
          },
          "rol_id": {
            "type": "integer",
            "title": "Rol Id"
          }
        },
        "type": "object",
        "required": [
          "consumo_id",
          "proyecto_id",
          "recurso_id",
          "fecha_inicio",
          "fecha_fin",
          "horas_consumidas",
          "tarea",
          "rol_id"
        ],
        "title": "ConsumoOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecursoIn": {
        "properties": {
          "recurso_nombre": {
            "type": "string",
            "title": "Recurso Nombre"
          },
          "es_admin": {
            "type": "boolean",
            "title": "Es Admin",
            "default": false
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "recurso_nombre"
        ],
        "title": "RecursoIn"
      },
...
      "UsuarioOut": {
        "properties": {
          "recurso_id": {
            "type": "integer",
            "title": "Recurso Id"
          },
          "recurso_nombre": {
            "type": "string",
            "title": "Recurso Nombre"
          },
          "es_admin": {
            "type": "boolean",
            "title": "Es Admin"
          },
          "debe_cambiar_password": {
            "type": "boolean",
            "title": "Debe Cambiar Password"
          }
        },
        "type": "object",
        "required": [
          "recurso_id",
          "recurso_nombre",
          "es_admin",
          "debe_cambiar_password"
        ],
        "title": "UsuarioOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}